requirements.txt
psycopg2-binary>=2.9.0  # PostgreSQL接続ライブラリの依存関係
